// DonutsboxClient/src/app/api/donutsbox/model/subscriptionPaymentResponseDto.ts
/**
 * Donutsbox.Api
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SubscriptionPaymentResponseDto { 
    /**
     * Идентификатор заявки на оплату в нашей системе.
     */
    paymentRequestId: string;
    /**
     * Идентификатор платежа в YooKassa.
     */
    paymentId: string | null;
    /**
     * Ссылка, на которую необходимо перенаправить пользователя для оплаты.
     */
    confirmationUrl: string | null;
    /**
     * Текущий статус платежа (pending/succeeded/canceled/...).
     */
    status: string | null;
}